Moving sensor data to models namespace
