+++ software/client/OpenDataDisc.Services/Interfaces/ISensorService.cs
@@ -1,4 +1,5 @@
 ﻿using System.Threading.Tasks;
+using OpenDataDisc.Services.Models;
 
 namespace OpenDataDisc.Services.Interfaces
 {
