+++ software/client/OpenDataDisc.Services/SensorService.cs
@@ -1,4 +1,5 @@
 ﻿using OpenDataDisc.Services.Interfaces;
+using OpenDataDisc.Services.Models;
 using System;
 using System.Data.SQLite;
 using System.Threading.Tasks;
